``` refactor: separate table state management from data fetching in founder table
- Refactored useTableData to accept pagination and search params instead of managing state internally
- Moved filter state (searchQuery) to parent component for better control
- Integrated useTableParams hook for pagination management before query execution
- Updated useTableState to consume params object and backend pagination data
- Removed redundant useEffect hooks that synced pagination state
- Simplified data
